ECMS-3562 Error on removing a new site

+++ core/services/src/main/java/org/exoplatform/services/wcm/category/RemoveTaxonomyPlugin.java
@@ -74,10 +74,12 @@ public class RemoveTaxonomyPlugin extends RemovePortalPlugin {
    */
   public void invalidateFromPortal(SessionProvider sessionProvider, String portalName) throws Exception {
     String repository = repositoryService.getCurrentRepository().getConfiguration().getName();
-    Node taxonomyTreeNode = taxonomyService.getTaxonomyTree(portalName, true);
-    if (taxonomyTreeNode!=null) {
-      actionServiceContainer.removeAction(taxonomyTreeNode, repository);
-    }
+    if (taxonomyService.hasTaxonomyTree(portalName)) {
+      Node taxonomyTreeNode = taxonomyService.getTaxonomyTree(portalName, true);
+      if (taxonomyTreeNode!=null) {
+        actionServiceContainer.removeAction(taxonomyTreeNode, repository);
+      }
       taxonomyService.removeTaxonomyTree(portalName);
+    }
   }
 }
